Better Magic link login form
