+++ apps/betterangels-backend/accounts/services.py
@@ -5,6 +5,7 @@ from django.conf import settings
 from django.core.mail import send_mail
 from django.template import loader
 from django.urls import reverse
+from post_office import mail
 from sesame.utils import get_query_string
 
 SUBJECT_LINE = "Your One-Click Login Link for the BetterAngels app"
@@ -24,12 +25,21 @@ def send_magic_link(email: str, base_url: str) -> str:
     text_template = loader.get_template("account/messages/email_magic_link.txt")
     text_message = text_template.render(context)
 
-    send_mail(
-        subject=subject,
-        message=text_message,
-        from_email=settings.DEFAULT_FROM_EMAIL,
-        recipient_list=[email],
-        html_message=html_message,
-    )
+    # mail.send(
+    #     [email],
+    #     settings.DEFAULT_FROM_EMAIL,
+    #     subject=subject,
+    #     message=text_message,
+    #     # html_message=html_message,
+    # )
+    # send_mail(
+    #     subject=subject,
+    #     message=text_message,
+    #     from_email="test@example.com",
+    #     recipient_list=["bloop@example.com"],
+    #     # html_message=html_message,
+    # )
+
+    send_mail(subject="test", message="tes2t", from_email="test@example.com", recipient_list=["admin@example.com"])
 
     return cast(str, query_string)
