+++ apps/betterangels-backend/betterangels_backend/apps.py
@@ -1,0 +1,5 @@
+from django.contrib.admin.apps import AdminConfig
+
+
+class MagicLinkAdminConfig(AdminConfig):
+    default_site = "betterangels_backend.admin.MagicLinkAdminSite"
